Fix issue when user enters a vertical bar character
When a user enters a vertical bar character in the task description, and it is saved in the file, when the file is opened again, it causes issues with the parsing of the file. Now, the user is no longer allowed to use the vertical bar character for the task description. User guide is updated to reflect this change.